+++ language/en/admin/colonies.php
@@ -1,0 +1,12 @@
+<?php
+
+$lang['title_moons'] = 'Moons list';
+$lang['title_planets'] = 'Planets list';
+
+$lang['col_name'] = 'Name';
+$lang['col_mother_planet'] = 'Mother planet';
+$lang['col_galaxy'] = 'Galaxy';
+$lang['col_system'] = 'System';
+$lang['col_planet'] = 'Planet';
+
+$lang['total'] = 'Total';